refactor(desktop): split reference workspace model
