+++ apps/desktop/src/features/library-workspace/references/model/folderAction.ts
@@ -1,0 +1,10 @@
+import type { ExplorerSnapshot } from '@/shared/types';
+
+export function getSelectableFolders(snapshot: ExplorerSnapshot) {
+  const statsByFolderId = new Map(snapshot.folderStats.map(stats => [stats.folderId, stats]));
+
+  return snapshot.virtualFolders.filter(folder => {
+    const stats = statsByFolderId.get(folder.id);
+    return folder.kind === 'user' && (stats?.childCount ?? 0) === 0;
+  });
+}
